feat: add goals authentication

+++ backend/controllers/goalController.js
@@ -1,13 +1,13 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
-
+const User = require("../models/userModel");
 /**
  * @desc Get goals
  * @route GET /api/goals
  * @access Private
  */
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -25,6 +25,7 @@ const setGoal = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(goal);
 });
@@ -41,6 +42,21 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found");
   }
+
+  const user = await User.findById(req.user.id);
+
+  // check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // make sure the logged in user matches the goal creator
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const updateGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -61,6 +77,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found");
   }
   try {
+    const user = await User.findById(req.user.id);
+
+    // check for user
+    if (!user) {
+      res.status(401);
+      throw new Error("User not found");
+    }
+
+    // make sure the logged in user matches the goal creator
+    if (goal.user.toString() !== user.id) {
+      res.status(401);
+      throw new Error("User not authorized");
+    }
     // remove()方法现在已经不用了
     await goal.deleteOne();
     res.status(200).json({ id: req.params.id });
